Time/date command: reorganize day and week of year function, also reoslving unbound local error. Fixes #34.
Unbound local error was caused by the Higri datetime being set to None if it is not an Islamic or Persian calendar, most prominent when trying to calculate remainig day count for the year. Therefore reorganize the entire function by:
1. Work at the calendar level - obtain all four tuple items (including days remaining).
2. Define the source list at the start of the function so it can be converted and returned as a tuple at the end.
This change introduces necessary code duplications due to change of algorithm.