Get day and week: partially revert and refine the last commit.
It turns out attempting to obtain persian year actualy slowed down Islamic year lookup because Gregorian to Persian date conversion took a while to complete (about 60 milliseconds or so). But not all is lost: have a variable which detects Islamic year and use it to define weekdays and appropriate calendar module.