Get day and week for Higri calendar: store month lengths inside a list.
Preivously convertdate.calendaer.month_length was called up to 23 times (up to 11 times to obtain day of the year, 12 times for total number of days for the given year). Therefore store month lengths inside a list of integers which wil then be used to:
* Calculate day of the year: sum the past month lengths, then add current date except for the first month where the day of the year is the current date.
* Calculate remaining days: sum of month lengths i.e. days for the year minus day of the year.
This reduces calendar.month_length call by half for worst case (if NVDA+F12 was pressed three times in December).