Get day and week: reconfigure the overall algorithm.
Since latter portion of calendar specific routines (adjusting the year and obtaining remainig days count) is the same across calendars, combine them into a single routine at the end. This means the role of the calendars will be to set up current day of the year, week, and total days for this calendar year, with the remaining day count calculated in the end. This, together with the other tuple components coming from different calendars, will be returned in the end.