Clock handler: define auto announce interval constants. Re #35
Use a combination of integer constatns and a dictionary to define clock auto announce intervals instead of using a tuple. This essentially means interval vlaue as specified in the add-on config will be looked up instead of coming up with interval tuples. This will be used in the clock handler class as part of report clock check - instead of a membership test, divmod will be used.