Clock handler: use a divmod when determining it is time to play chime and/or speech output. Re #35
Instead of a membership test, use divmod(minute, interval) to determine if it is time to play the chime and/or announce time via speech. Note that attempting to do divmod(minute, 0) results in zero divisoin error (that's why auto nanounce interval function returns 0 if interval is not set), and the second member of the divmod tuple holds modulo which must be 0 (interval has arrived) before checking seconds.