Add-on settings/quiet hours: replace 'input in 24-hour format' checkbox with 'quiet hours time format' combo box.
Because combo boxes will be used to specify quiet hours, a combo box to select hour format would be appropriate (besudes, in previous add-on releases, people weren't sure as to the entry format to be used for hours). Thereforr replace 'input in 24-hour format' checkbox with 'quiet hours time format' combo box with '12-hour' and '24-hour' as options. Another purpose of this new combo box is to change hours combo box item display depending on hour format chosen (in a future commit).