Add-on settings/quiet hours: remove time format check as it is always correct.
Part of the reason for replacing quiet hours text fields with combo boxes: to remove format checks (and what appears to be not a really helpful error message which insists users should read the documentation). With the time strings parsed to appropriate format, there is no need for format checks.