Schedule alarms: guard against multiple instances (try/except).
Suppose schedule alarms dialog is open and the user wishes to open another instance of this dialog. If so, NVDA will throw a multi-instance error because schedule alarms is a settings dialog. Therefore guard against this by enclosing the open method inside a try block and present an error if the user does want to open schedule alarms dialog again.