Datetime/hour range: use a series of 'or' statements when computing time range.
Huge optimization/assumption change: assume 24-hour range within the same day. For example, if start time is 12:00 and end time is 06:00, assume 06:00 as in earlier that day (effectivley, this becomes 06:00 to 12:00). This then allows a series of 'or' statements to be executed (later than start time or earlier than end time), assuming that this is a 24-hour range. The practical effect is that there is no need for a one day time delta anymore, simplifying time range function significantly.