New time formats and ability to separate interval and hourly chimes
This commit introduces new time formats ($$H $$mm $$ss and $$HH $$mm $$ss) which read hours, minutes and seconds without colons and new settings that allow users to customize the sounds for both interval and hourly chimes and a checkbox to enable this feature (off by default. An Italian translation for the new settings is provided.